move octetsplitfunc to syslog input
