+++ pkg/stanza/operator/input/syslog/syslog.go
@@ -4,8 +4,11 @@
 package syslog // import "github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator/input/syslog"
 
 import (
+	"bufio"
 	"errors"
 	"fmt"
+	"regexp"
+	"strconv"
 
 	"go.uber.org/zap"
 
@@ -58,9 +61,11 @@ func (c Config) Build(logger *zap.SugaredLogger) (operator.Operator, error) {
 
 	if c.TCP != nil {
 		tcpInputCfg := tcp.NewConfigWithID(inputBase.ID() + "_internal_tcp")
-		c.TCP.Multiline.OctetCounting = syslogParserCfg.EnableOctetCounting
-		tcpInputCfg.BaseConfig = *c.TCP
+		if syslogParserCfg.EnableOctetCounting {
+			tcpInputCfg.MultiLineBuilder = OctetMultiLineBuilder
+		}
 
+		tcpInputCfg.BaseConfig = *c.TCP
 		tcpInput, err := tcpInputCfg.Build(logger)
 		if err != nil {
 			return nil, fmt.Errorf("failed to resolve tcp config: %w", err)
@@ -136,4 +141,41 @@ func (t *Input) Stop() error {
 func (t *Input) SetOutputs(operators []operator.Operator) error {
 	t.parser.SetOutputIDs(t.GetOutputIDs())
 	return t.parser.SetOutputs(operators)
+}
+
+func OctetMultiLineBuilder() (bufio.SplitFunc, error) {
+	return newOctetFrameSplitFunc(true), nil
+}
+
+func newOctetFrameSplitFunc(flushAtEOF bool) bufio.SplitFunc {
+	frameRegex := regexp.MustCompile(`^[1-9]\d*\s`)
+	return func(data []byte, atEOF bool) (advance int, token []byte, err error) {
+		frameLoc := frameRegex.FindIndex(data)
+		if frameLoc == nil {
+			// Flush if no more data is expected
+			if len(data) != 0 && atEOF && flushAtEOF {
+				token = data
+				advance = len(data)
+				return
+			}
+			return 0, nil, nil
+		}
+
+		frameMaxIndex := frameLoc[1]
+		// delimit space between length and log
+		frameLenValue, err := strconv.Atoi(string(data[:frameMaxIndex-1]))
+		if err != nil {
+			return 0, nil, err // read more data and try again.
+		}
+
+		advance = frameMaxIndex + frameLenValue
+		// the limitation here is that we can only line split within a single buffer
+		// the context of buffer length cannot be pass onto the next scan
+		if advance > cap(data) {
+			return 0, nil, errors.New("frame size is larger than buffer capacity")
+		}
+		token = data[:advance]
+		err = nil
+		return
+	}
 }
